segregate interfaces
