+++ app/Structures/GuestUserLessonRepositoryInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Structures;
+
+use Illuminate\Support\Collection;
+
+interface GuestUserLessonRepositoryInterface
+{
+    /**
+     * @return Collection
+     */
+    public function fetchPublicUserLessons(): Collection;
+}
