+++ app/Structures/UserLessonRepositoryInterface.php
@@ -3,7 +3,6 @@
 namespace App\Structures;
 
 use App\Models\User;
-use Illuminate\Support\Collection;
 
 interface UserLessonRepositoryInterface
 {
@@ -13,27 +12,4 @@ interface UserLessonRepositoryInterface
      * @return UserLesson|null
      */
     public function fetchUserLesson(?User $user, int $lessonId): ?UserLesson;
-
-    /**
-     * @param User $user
-     * @return Collection
-     */
-    public function fetchOwnedUserLessons(User $user): Collection;
-
-    /**
-     * @param User $user
-     * @return Collection
-     */
-    public function fetchSubscribedUserLessons(User $user): Collection;
-
-    /**
-     * @param User $user
-     * @return Collection
-     */
-    public function fetchAvailableUserLessons(User $user): Collection;
-
-    /**
-     * @return Collection
-     */
-    public function fetchPublicUserLessons(): Collection;
 }
